HTMLInputElement::setEditingValue should not fail if renderer doesn't exist https://bugs.webkit.org/show_bug.cgi?id=195708
Reviewed by Wenson Hsieh.

Source/WebCore:

HTMLInputElement::setEditingValue currently returns early if the element's renderer() is
null. This is causing the Epiphany password manager to fail to remember passwords on
https://www.geico.com/ except for navigations through page cache.

This check was originally added to avoid some assertion, but I don't know which one, and
there's definitely not any assertion hit nowadays in this case. Probably there are more
guards checking if renderer() is null elsewhere in the code nowadays, closer to where it's
really needed.

Test: fast/forms/editing-value-null-renderer.html

* html/HTMLInputElement.cpp:
(WebCore::HTMLInputElement::setEditingValue):

LayoutTests:

This is a copy of fast/forms/editing-value.html, except the form is not displayed. The input
value should still change.

* fast/forms/editing-value-null-renderer-expected.txt: Added.
* fast/forms/editing-value-null-renderer.html: Added.


Identifier: 210667@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243647 268f45cc-cd09-0410-ab3c-d52691b4dbfc