REGRESSION (r243523): Six form-related watchOS layout tests are timing out https://bugs.webkit.org/show_bug.cgi?id=196405 <rdar://problem/49428130>
Reviewed by Wenson Hsieh.

Tools:

Add a shouldPresentPopovers=false WebKitTestRunner option to cause WKTR to swizzle
the popover presentation methods to be no-ops. Use this in the new test added in
r243523 so that we do not swizzle those methods for all tests. This is needed because
those WatchOS tests rely on the popover getting presented and interactive.

* WebKitTestRunner/TestController.cpp:
(WTR::updateTestOptionsFromTestHeader):
* WebKitTestRunner/TestController.h:
* WebKitTestRunner/TestOptions.h:
(WTR::TestOptions::hasSameInitializationOptions const):
* WebKitTestRunner/ios/TestControllerIOS.mm:
(WTR::TestController::platformInitialize):
(WTR::TestController::platformResetStateToConsistentValues):

LayoutTests:

* fast/forms/ios/file-upload-panel.html:


Identifier: 210674@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243657 268f45cc-cd09-0410-ab3c-d52691b4dbfc