Assertion failed in JSC::createError https://bugs.webkit.org/show_bug.cgi?id=196305 <rdar://problem/49387382>
Reviewed by Saam Barati.

JSTests:

* stress/create-error-out-of-memory-rope-string-2.js: Added.
(assert):
(catch):

Source/JavaScriptCore:

JSC::createError assumes that `errorDescriptionForValue` will either
throw an exception or return a valid description string. However, that
is not true if the value is a rope string and we successfully resolve it,
but later fail to wrap the string in quotes with `tryMakeString`.

* runtime/ExceptionHelpers.cpp:
(JSC::createError):


Identifier: 210682@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@243665 268f45cc-cd09-0410-ab3c-d52691b4dbfc