Ensure iOS layout traits are used for media controls in modern compatibility mode https://bugs.webkit.org/show_bug.cgi?id=196812 <rdar://problem/47460637>
Reviewed by Dean Jackson.

Source/WebCore:

Test: media/modern-media-controls/media-controller/ios/media-controller-ios-layout-traits-modern-compatibility-mode.html

Instead of looking at the UA string, check whether we support touches which is the correct indicator of whether we should
be using the iOS layout traits for media controls.

* Modules/modern-media-controls/media/media-controller.js:
(MediaController.prototype.get layoutTraits):

LayoutTests:

Add a new test that enforces modern compatibility mode and checks that the iOS controls are used.

* media/modern-media-controls/media-controller/ios/media-controller-ios-layout-traits-modern-compatibility-mode-expected.txt: Added.
* media/modern-media-controls/media-controller/ios/media-controller-ios-layout-traits-modern-compatibility-mode.html: Added.

Identifier: 211160@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244253 268f45cc-cd09-0410-ab3c-d52691b4dbfc