Web Inspector: InspectorFrontendHost::inspectionLevel returns wrong value for inspector3 https://bugs.webkit.org/show_bug.cgi?id=197044
Reviewed by Joseph Pecoraro.

* UIProcess/WebInspectorUtilities.h:
* UIProcess/WebInspectorUtilities.cpp:
(WebKit::trackInspectorPage):
(WebKit::untrackInspectorPage):
Determine the level of the inspector page by looking to see if the page being inspected is
in the inspector page map. If so, we must be inspecting an inspector page and should use
that inspector page's inspection level as the value to increment.

* UIProcess/WebInspectorProxy.cpp:
(WebKit::WebInspectorProxy::createFrontendPage):
* UIProcess/RemoteWebInspectorProxy.cpp:
(WebKit::RemoteWebInspectorProxy::createFrontendPageAndWindow):


Identifier: 211302@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244418 268f45cc-cd09-0410-ab3c-d52691b4dbfc